More schema updates

+++ PhraserManager/src/main/java/com/phraser/db/PhraseBlock.java
@@ -10,7 +10,7 @@ public interface PhraseBlock extends StoreBlock {
     interface Word {
         /** 16 bit */
         int wordTemplateId();
-        int wordTemplateOrdinal();
+        short wordTemplateOrdinal();
         String name();
         String word();
         byte permissions();
